Move PyVista code from vtk_core to new vtk_utilities
