+++ pyproject.toml
@@ -495,7 +495,8 @@ allow-dict-calls-with-keyword-arguments = true
 'pyvista/core/_typing_core/_dataset_types.py' = ['PYI018']
 'pyvista/core/_validation/_cast_array.py' = ['ANN']
 'pyvista/core/_validation/validate.py' = ['ANN202']
-'pyvista/core/_vtk_core.py' = ['ANN', 'PLC0414']
+'pyvista/core/_vtk_core.py' = ['PLC0414']
+'pyvista/core/_vtk_utilities.py' = ['ANN']
 'pyvista/core/dataobject.py' = ['PLC0415']
 'pyvista/core/errors.py' = ['ANN']
 'pyvista/core/filters/*' = ['ANN', 'PYI019']
