+++ pyvista/core/cell.py
@@ -9,6 +9,8 @@ import numpy as np
 
 import pyvista as pv
 from pyvista._deprecate_positional_args import _deprecate_positional_args
+from pyvista.core._vtk_utilities import DisableVtkSnakeCase
+from pyvista.core._vtk_utilities import vtkPyVistaOverride
 
 from . import _vtk_core as _vtk
 from ._typing_core import BoundsTuple
@@ -619,8 +621,8 @@ class Cell(_BoundsSizeMixin, DataObject, _vtk.vtkGenericCell):
 
 class CellArray(
     _NoNewAttrMixin,
-    _vtk.DisableVtkSnakeCase,
-    _vtk.vtkPyVistaOverride,
+    DisableVtkSnakeCase,
+    vtkPyVistaOverride,
     _vtk.vtkCellArray,
 ):
     """PyVista wrapping of :vtk:`vtkCellArray`.
