+++ pyvista/core/composite.py
@@ -28,6 +28,7 @@ from typing_extensions import Unpack
 import pyvista as pv
 from pyvista._deprecate_positional_args import _deprecate_positional_args
 from pyvista.core import _validation
+from pyvista.core._vtk_utilities import vtk_version_info
 
 from . import _vtk_core as _vtk
 from ._typing_core import BoundsTuple
@@ -151,7 +152,7 @@ class MultiBlock(
     _WRITERS: ClassVar[dict[str, type[BaseWriter]]] = dict.fromkeys(
         ['.vtm', '.vtmb'], XMLMultiBlockDataWriter
     )
-    if _vtk.vtk_version_info >= (9, 4):
+    if vtk_version_info >= (9, 4):
         _WRITERS['.vtkhdf'] = HDFWriter
 
     def __init__(
