+++ pyvista/core/utilities/reader.py
@@ -25,7 +25,7 @@ import numpy as np
 import pyvista as pv
 from pyvista._deprecate_positional_args import _deprecate_positional_args
 from pyvista.core import _vtk_core as _vtk
-from pyvista.core._vtk_core import VersionInfo
+from pyvista.core._vtk_utilities import VersionInfo
 
 from .fileio import _FileIOBase
 from .fileio import _get_ext_force
