+++ pyvista/plotting/axes.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 from pyvista._deprecate_positional_args import _deprecate_positional_args
+from pyvista.core._vtk_utilities import DisableVtkSnakeCase
 from pyvista.core.utilities.misc import _NoNewAttrMixin
 
 from . import _vtk
@@ -10,7 +11,7 @@ from .actor import Actor
 from .axes_actor import AxesActor
 
 
-class Axes(_NoNewAttrMixin, _vtk.DisableVtkSnakeCase, _vtk.vtkAxes):
+class Axes(_NoNewAttrMixin, DisableVtkSnakeCase, _vtk.vtkAxes):
     """PyVista wrapper for the VTK Axes class.
 
     Parameters
