+++ pyvista/plotting/composite_mapper.py
@@ -12,6 +12,7 @@ import numpy as np
 import pyvista as pv
 from pyvista import vtk_version_info
 from pyvista._deprecate_positional_args import _deprecate_positional_args
+from pyvista.core._vtk_utilities import DisableVtkSnakeCase
 from pyvista.core.utilities.arrays import convert_array
 from pyvista.core.utilities.arrays import convert_string_array
 from pyvista.core.utilities.misc import _check_range
@@ -265,7 +266,7 @@ class BlockAttributes(_NoNewAttrMixin):
 
 class CompositeAttributes(
     _NoNewAttrMixin,
-    _vtk.DisableVtkSnakeCase,
+    DisableVtkSnakeCase,
     _vtk.vtkCompositeDataDisplayAttributes,
 ):
     """Block attributes.
